add createContainer for UserCard in order to get skills to display
