+++ imports/ui/views/organizations/OrganizationCard/OrganizationCardContainer.jsx
@@ -1,7 +1,6 @@
 import { Meteor } from 'meteor/meteor'
 import { createContainer } from 'meteor/react-meteor-data'
 import React, { Component } from 'react'
-import _ from 'lodash'
 
 import { Organizations } from '/imports/api/Organizations'
 import { Tags } from '/imports/api/Tags'
